refactor(showcase): to ts
